+++ tests/Remote/LoggerTest.php
@@ -14,7 +14,7 @@ use Innmind\IO\Sockets\{
     Clients\Client,
 };
 use Innmind\Url\Url;
-use Innmind\Immutable\Maybe;
+use Innmind\Immutable\Attempt;
 use Formal\AccessLayer\Connection;
 use Psr\Log\LoggerInterface;
 use PHPUnit\Framework\TestCase;
@@ -82,7 +82,7 @@ class LoggerTest extends TestCase
                     ->expects($this->once())
                     ->method('socket')
                     ->with($transport, $authority)
-                    ->willReturn($expected = Maybe::just(null /* hack to avoid creating real client */));
+                    ->willReturn($expected = Attempt::result(null /* hack to avoid creating real client */));
                 $logger = $this->createMock(LoggerInterface::class);
                 $logger
                     ->expects($this->once())